nand_ecc: Align the nand 1bit correction ECC routines to linux v3.7
This patch aligns the nand_calculate_ecc() and nand_correct_data()
routines used to correct 1 error bit on a block of data to the
linux v3.7. In particular following linux patches are now included:

commit e6cf5df1838c28bb060ac45b5585e48e71bbc740
Author: frans <fransmeulenbroeks@gmail.com>
Date:   Fri Aug 15 23:14:31 2008 +0200

    [MTD] [NAND] nand_ecc.c: rewrite for improved performance

commit d68156cfad0fe09201dd049fff167a8a881427ad
Author: Singh, Vimal <vimalsingh@ti.com>
Date:   Sat Aug 23 18:18:34 2008 +0200

    [MTD] [NAND] nand_ecc.c: adding support for 512 byte ecc

Please note that without this changes platforms that declare
ecc.size >= 512 bytes are not able to correct error bits in the proper
way.

Signed-off-by: Armando Visconti <armando.visconti@st.com>